Integration of error trapping functions and addition of search engine endpoints

+++ database/migrations/2024_08_15_141805_create_employees_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('employees', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('company_id');
+            $table->unsignedInteger('company_id');
             $table->string('first_name', 255);
             $table->string('last_name', 255);
             $table->string('email', 255)->unique();
